fix #24
I've removed the single offending line of code and everything seems to
be functioning the way it was before. The View::subscribe method is deprecated
and seems to have been the cause of the problem. I'm not entirely sure
how disposables work, but replacing this line of code does not seem to
be nessecary.