Updated atp-view.coffee
moveToCurrentDirectory in atp-view.coffee was not working on my mac so
I found the culprit and wrote fixed the script. I’m assuming that it
will work universally now but I could be mistaken. Someone will need to
test it on Windows.

The problem was the regex in the getCurrentFileName not matching the
file path correctly. Fixed it by getting the file object from the
activeTextEditor and grabbing the base name. Works correctly now on my
mac.